% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{add}
\alias{add}
\title{Add Documents}
\arguments{
\item{doc}{tibble of which the first column represents document id
and the second column represents document body.}
}
\description{
Add documents to index.
}
