% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conifer.R
\docType{package}
\name{conifer}
\alias{conifer}
\alias{conifer-package}
\title{conifer: R client of COTOHA API for Developers}
\description{
R client of COTOHA API for Developers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/conifer}
  \item Report bugs at \url{https://github.com/paithiov909/conifer/issues}
}

}
\author{
\strong{Maintainer}: Kato Akiru \email{paithiov909@gmail.com}

}
