% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{coreference}
\alias{coreference}
\title{Reference Resolution}
\arguments{
\item{sentence}{string.}

\item{type}{string. Select `default` or `kuzure`.}
}
\value{
list.
}
\description{
Reference Resolution
}
