% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotoha.R
\name{cotoha}
\alias{cotoha}
\title{COTOHA API Client}
\usage{
cotoha(access_token, base_url = "https://api.ce-cotoha.com/api/dev")
}
\arguments{
\item{access_token}{string.}

\item{base_url}{string.}
}
\value{
R6 class object having methods below.
\itemize{
  \item parse(): Parsing dependency structure
  \item ne(): Named entity extraction
  \item coreference(): Reference resolution
  \item keyword(): Keyword extraction
  \item similarity(): Calculating similality between sentences
  \item sentence_type(): Linguistic modality detection
  \item user_attribute(): Speaker's characteristics recognition
  \item remove_filler(): Removing fillers of sentence
  \item detect_misrecognition(): Detecting misrecognition of sentence
  \item sentiment(): Sentiment analysis
}
}
\description{
Return client class object.
}
\seealso{
\url{https://api.ce-cotoha.com/contents/reference/apireference.html}
}
