% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{detect_misrecognition}
\alias{detect_misrecognition}
\title{Detecting Misrecognition of Sentence}
\arguments{
\item{sentence}{string.}
}
\value{
list.
}
\description{
Detecting Misrecognition of Sentence
}
