% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{documentTermMatrix}
\alias{documentTermMatrix}
\title{Create Document-Term-Matrix}
\arguments{
\item{dtm_name}{string. variable name that the DTM will be assigned.}

\item{percent}{integer. Function removes all words from the DTM which appear in less than `percent` of the documents.}

\item{weight}{boolean. If true, the DTM is weighted with TF-IDF.}
}
\description{
This function assigns documemt-term-matrix (DTM) to variable on R environment.
Each DTM is a list that has properties below.
\itemize{
  \item vocabulary
  \item data
  \item nDocs
  \item nTerms
}
}
\seealso{
\url{https://github.com/Planeshifter/text-miner#documenttermmatrix--termdocumentmatrix}
}
