% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticlunr.R
\name{elasticlunr}
\alias{elasticlunr}
\title{Create Documents Index with elasticlunr.js}
\usage{
elasticlunr(ctx = NULL)
}
\arguments{
\item{ctx}{Your own V8 context if any.}
}
\value{
R6 class object having methods below.
\itemize{
  \item add(): Add documents to index.
  \item remove(): Remove documents from index.
  \item update(): Update documents in index.
  \item search(): Search document of index.
}
}
\description{
Return an elasticlunr.js V8 binding.
}
