% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{expandContractions}
\alias{expandContractions}
\title{Replace English Contractions}
\description{
Replace all occuring English contractions by their expanded equivalents, e.g. "don't" is changed to "do not".
}
