% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotoha.R
\name{getAccessToken}
\alias{getAccessToken}
\title{Getting Access Token of COTOHA API}
\usage{
getAccessToken(
  publish_url,
  client_id = Sys.getenv("COTOHA_ID"),
  client_secret = Sys.getenv("COTOHA_SECRET"),
  grant_type = "client_credentials"
)
}
\arguments{
\item{publish_url}{string.}

\item{client_id}{string.}

\item{client_secret}{string.}

\item{grant_type}{string.}
}
\value{
return access token only when the authentication successed. If failed, then return httr response invisibly.
}
\description{
Getting Access Token of COTOHA API
}
\seealso{
\url{https://api.ce-cotoha.com/contents/reference/accesstoken.html}
}
