% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conifer-reference.R
\name{keyword}
\alias{keyword}
\title{Keyword Extraction}
\arguments{
\item{documetn}{character.}

\item{type}{string. Select `default` or `kuzure`.}

\item{do_segment}{boolean. Default is false.}
}
\value{
list.
}
\description{
Keyword Extraction
}
