% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noramalizer.R
\name{normalize}
\alias{normalize}
\title{Normalize Text on V8 Context}
\usage{
normalize(str)
}
\arguments{
\item{str}{string.}
}
\value{
normalized text.
}
\description{
Normalize Text on V8 Context
}
\seealso{
\url{https://github.com/moznion/neologd-normalizer-js}
}
