% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{parse}
\alias{parse}
\title{Parsing Dependency Structure}
\arguments{
\item{sentence}{string.}

\item{type}{string. Select `default` or `kuzure`.}
}
\value{
list.
}
\description{
Parsing Dependency Structure
}
