% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pretty_chunks}
\alias{pretty_chunks}
\title{Preffy Parsed Chunks}
\usage{
pretty_chunks(li)
}
\arguments{
\item{li}{List. Output of \code{client$parse()}.}
}
\value{
tibble.
}
\description{
Preffy Parsed Chunks
}
