% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conifer-reference.R
\name{remove_filler}
\alias{remove_filler}
\title{Removing Fillers of Sentence}
\arguments{
\item{text}{string.}

\item{do_segment}{boolean. Default is false.}
}
\value{
list.
}
\description{
Removing Fillers of Sentence
}
