% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{removeWords}
\alias{removeWords}
\title{Remove Supplied Words}
\arguments{
\item{stop_words}{characters.}

\item{case_insensitive}{boolean. Default is false.}
}
\description{
Remove all words in the supplied `stop_words` array from all documents.
}
