% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{search}
\alias{search}
\title{Search Document}
\arguments{
\item{query}{List of query strings.}

\item{expand}{boolean. If true, it will expand a query token to increase recall.
Default is FALSE.}

\item{boolean}{Boolean logic setting string. "OR" or "AND" is acceptable. Default is "OR".}

\item{boost}{Used for Query-Time Boosting setting.
So that each documents of the index has only one value field, you do not have to change this value.}
}
\description{
Search document of index.
}
