% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conifer-reference.R
\name{sentence_type}
\alias{sentence_type}
\title{Linguistic Modality Detection}
\arguments{
\item{sentence}{string.}

\item{type}{string. Select `default` or `kuzure`.}
}
\description{
Linguistic Modality Detection
}
