% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{similarity}
\alias{similarity}
\title{Calculating Similality between Sentences}
\arguments{
\item{s1}{string.}

\item{s2}{string.}

\item{type}{string. Select `default` or `kuzure`.}
}
\value{
list.
}
\description{
Calculating Similality between Sentences
}
