% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{stem}
\alias{stem}
\title{Stemming of the Words}
\arguments{
\item{type}{string. Select `Porter` or `Lancaster`. Default is `Porter`.}
}
\description{
Perform stemming of the words in each document. Two stemmers are supported: Porter and Lancaster.
}
