% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textminer.R
\name{text_miner}
\alias{text_miner}
\title{Create Corpus with text-miner}
\usage{
text_miner(ctx = NULL)
}
\arguments{
\item{ctx}{Your own V8 context if any.}
}
\value{
R6 class object having methods below.
\itemize{
  \item addDoc()
  \item addDocs()
  \item clean()
  \item removeInterpunctuation()
  \item removeNewlines()
  \item removeWords()
  \item removeDigits()
  \item removeInvalidCharacters()
  \item stem()
  \item toLower()
  \item toUpper()
  \item trim()
  \item documentTermMatrix()
  \item expandContractions()
}
}
\description{
Return a Planeshifter/text-miner V8 binding.
}
\seealso{
\url{https://github.com/Planeshifter/text-miner}
}
