% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{update}
\alias{update}
\title{Update Documents}
\arguments{
\item{doc}{tibble of which the first column represents document id
and the second column represents document body.}
}
\description{
Update documents in index.
}
