% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conifer-reference.R
\name{user_attribute}
\alias{user_attribute}
\title{Speaker's Characteristics Recognition}
\arguments{
\item{document}{character.}

\item{do_segment}{boolean. Default is false.}
}
\value{
list.
}
\description{
Speaker's Characteristics Recognition
}
