% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yamanashi}
\alias{yamanashi}
\title{Yamanashi}
\format{
An object of class \code{character} of length 87.
}
\source{
\url{https://www.aozora.gr.jp/cards/000081/card46605.html}
}
\usage{
yamanashi
}
\description{
UTF-8 text data of `Yamanashi` written by Miyazawa Kenji.
}
\keyword{datasets}
