% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conifer-package.R
\docType{package}
\name{conifer}
\alias{conifer}
\alias{conifer-package}
\title{conifer: R client of COTOHA API for Developers}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

R client of COTOHA API for Developers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/conifer}
  \item Report bugs at \url{https://github.com/paithiov909/conifer/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
