% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conifer-reference.R
\name{coreference}
\alias{coreference}
\title{Reference Resolution}
\arguments{
\item{document}{character.}

\item{type}{string. Select `default` or `kuzure`.}

\item{do_segment}{boolean. Default is false.}
}
\value{
list.
}
\description{
Reference Resolution
}
